fix: add error handling for inactive farms and unauthorized access in delete farm image use case

+++ src/core/use-cases/delete-farm-image.ts
@@ -6,8 +6,9 @@ import { UsersRepository } from "@/core/repositories/users-repository";
 import { Storage } from "@/core/storage/storage";
 
 // Errors
+import { FarmNotActiveError } from "@/core/errors/farm-not-active";
 import { ResourceNotFoundError } from "@/core/errors/resource-not-found";
-
+import { UnauthorizedError } from "@/core/errors/unauthorized";
 interface DeleteFarmImageUseCaseRequest {
   user_id: string;
   farm_id: string;
@@ -30,16 +31,18 @@ export class DeleteFarmImageUseCase {
 
     if (!farm) throw new ResourceNotFoundError("Fazenda", farm_id);
 
-    const image = farm.images.get(image_url);
-
-    if (!image) throw new ResourceNotFoundError("Imagem", image_url);
-
     const user = await this.usersRepository.find("basic", { id: user_id });
 
     if (!user) throw new ResourceNotFoundError("Usuário", user_id);
 
     if (!farm.admin_id.equals(user.id) && !user.admin)
-      throw new ResourceNotFoundError("Fazenda", farm_id);
+      throw new UnauthorizedError();
+
+    if (farm.status !== "ACTIVE") throw new FarmNotActiveError();
+
+    const image = farm.images.get(image_url);
+
+    if (!image) throw new ResourceNotFoundError("Imagem", image_url);
 
     await this.storage.delete(image, "farms");
